Move InMemoryTree into memory subpackage
